requirements.txt
python>=3.8
torch>=2.0
transformers>=4.38.0
accelerate>=0.26.0
sentence-transformers>=2.6.0
Pillow>=9.5.0
numpy>=1.23.0
chromadb>=0.4.24
pypdf>=3.9.0
gradio>=4.25.0
tqdm>=4.65.0
regex>=2023.10.3
